Move ResolveInfo to Execution namespace
