+++ src/Execution/ExecutionStrategy.php
@@ -6,7 +6,7 @@ use Digia\GraphQL\Error\ExecutionException;
 use Digia\GraphQL\Error\GraphQLException;
 use Digia\GraphQL\Error\InvalidTypeException;
 use Digia\GraphQL\Error\UndefinedException;
-use Digia\GraphQL\Execution\Resolver\ResolveInfo;
+use Digia\GraphQL\Execution\ResolveInfo;
 use Digia\GraphQL\Language\Node\FieldNode;
 use Digia\GraphQL\Language\Node\FragmentDefinitionNode;
 use Digia\GraphQL\Language\Node\FragmentSpreadNode;
